+++ src/index.ts
@@ -1,11 +1,12 @@
 export { cn } from "./lib/utils";
 export { Button } from "./components/ui/button";
 export { buttonVariants } from "./components/ui/button.variants";
-export { Badge } from './components/ui/badge'
-export { badgeVariants } from './components/ui/badge.variants'
+export { Badge } from "./components/ui/badge";
+export { badgeVariants } from "./components/ui/badge.variants";
 export { Input } from "./components/ui/input";
 export { Label } from "./components/ui/label";
 export { Separator } from "./components/ui/separator";
+export { Progress } from "./components/ui/progress";
 export {
   Field,
   FieldLabel,